Updated the read from /etc/passwd
In /etc/passwd, there are many users including system users that are not relevant to us. The read from /etc/passwd therefore also prompted for system users. This was time-consuming, as there are more than 50 of them. So, I added an awk command before the cut to filter out only the regular users with an expression that only allowed users with a /home directory.